Place APIs under http package
