+++ modules/examples/todolist-http-finch/src/main/scala/todo/http/CRUDApi.scala
@@ -16,32 +16,17 @@
 
 package examples.todolist
 package http
-package apis
 
-/**
- * Finch http endpoints
- */
-class Api[F[_]](
-    implicit appApi: AppApi[F],
-    genericApi: GenericApi[F],
-    todoItemApi: TodoItemApi[F],
-    todoListApi: TodoListApi[F],
-    tagApi: TagApi[F]) {
+import io.finch.Endpoint
 
-  val endpoints =
-    appApi.endpoints :+:
-      genericApi.endpoints :+:
-      todoItemApi.endpoints :+:
-      todoListApi.endpoints :+:
-      tagApi.endpoints
+trait CRUDApi[A] {
 
-}
+  val reset: Endpoint[Int]
+  val retrieve: Endpoint[A]
+  val list: Endpoint[List[A]]
+  val insert: Endpoint[Option[A]]
+  val update: Endpoint[Option[A]]
+  val destroy: Endpoint[Int]
 
-object Api {
-  implicit def instance[F[_]](
-      implicit appApi: AppApi[F],
-      genericApi: GenericApi[F],
-      todoItemApi: TodoItemApi[F],
-      todoListApi: TodoListApi[F],
-      tagApi: TagApi[F]): Api[F] = new Api[F]
+  lazy val endpoints = reset :+: retrieve :+: list :+: insert :+: update :+: destroy
 }
